requirements.txt
annotated-types==0.7.0
anyio==4.9.0
dnspython==2.7.0
fastapi==0.115.12
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
